Add `git push origin gh-pages` in main repo.
The original makefile did a `git push` down in pages/, however that
gives me a warning with newer git, so I changed it to `git push origin
gh-pages`. However I _think_ that meant that it was simply updating the
main repo from the git clone repo down in pages/. As a result, things
weren't actually getting pushed up to GitHub.

So I added another `git push origin gh-pages`, from the main repo, as
the final step.

I don't know if this is the optimal way to resolve this, but it's the
way that works for me based on my understanding (perhaps poor) of the
relationship of the pages/ repo to the main one.