Simplify `parse`: Always make rename transformer.
The rename transformer makes `%` an alias for `%1` (which is a really
smart way to handle this!).

Prior to this commit, `parse` avoided creating the rename transformer in
the case where the function has no arguments at all, IIUC because it
wouldn't be used.

However more generally, `%` won't be used if `body` doesn't
use `%` -- which wasn't being checked for.

If defining the rename transformer when it won't be used, is "wasteful",
we should really check for the general case.

However. If I understand correctly, it's not really "wasteful". So let's
not bother checking at all. Let's simply define it, always.